Add SVG icons for dental treatments and update treatments list
- Added SVG icons for various dental treatments including:
  - Teeth Whitening
  - Emergency Dental Care
  - Dental Exam & X-Rays
  - Hygiene Services & Dental Cleaning
  - Dental Implants
  - Tooth Fillings
  - Invisalign
  - Partial & Full Dentures
  - Root Canals

- Updated treatments.jsx to include new icons and treatment names.
