--- src/assets/treatments.js
@@ -1,38 +1,0 @@
-export const treatments = [
-  {
-    name: 'Teeth Whitening',
-    img: 'https://lumino.co.nz/media/2008/teeth-whitening.svg?width=200&v=1da27015d3d4d00'
-  },
-  {
-    name: 'Emergency Dental Care',
-    img: 'https://lumino.co.nz/media/2009/emergency-dental-care.svg?width=200&v=1da27015d3d4d00'
-  },
-  {
-    name: 'Dental Exam & X-Rays',
-    img: 'https://lumino.co.nz/media/2010/dental-exam-x-rays.svg?width=200&v=1da27015d3d4d00'
-  },
-  {
-    name: 'Hygiene Services & Dental Cleaning',
-    img: 'https://lumino.co.nz/media/2011/hygiene-services-dental-cleaning.svg?width=200&v=1da27015d3d4d00'
-  },
-  {
-    name: 'Dental Implants',
-    img: 'https://lumino.co.nz/media/2012/dental-implants.svg?width=200&v=1da27015d3d4d00'
-  },
-  {
-    name: 'Tooth Fillings',
-    img: 'https://lumino.co.nz/media/2013/tooth-fillings.svg?width=200&v=1da27015d3d4d00'
-  },
-  {
-    name: 'Invisalign®',
-    img: 'https://lumino.co.nz/media/2014/invisalign.svg?width=200&v=1da27015d3d4d00'
-  },
-  {
-    name: 'Partial & Full Dentures',
-    img: 'https://lumino.co.nz/media/2015/partial-full-dentures.svg?width=200&v=1da27015d3d4d00'
-  },
-  {
-    name: 'Root Canals',
-    img: 'https://lumino.co.nz/media/2016/root-canals.svg?width=200&v=1da27015d3d4d00'
-  }
-];
