+++ src/assets/treatments.jsx
@@ -1,0 +1,48 @@
+import teethWhitening from './images/teeth-whitening.svg';
+import emergencyDentalCare from './images/emergency-icon.svg';
+import dentalExamXRays from './images/x-rays.svg';
+import hygieneServicesDentalCleaning from './images/dental-hygiene.svg';
+import dentalImplants from './images/dental-implant.png';
+import toothFillings from './images/fillings.svg';
+import invisalign from './images/invisalign.svg';
+import partialFullDentures from './images/dentures.svg';
+import rootCanals from './images/endodontiscs.svg';
+
+export const treatments = [
+  {
+    name: 'Teeth Whitening',
+    img: teethWhitening
+  },
+  {
+    name: 'Emergency Dental Care',
+    img: emergencyDentalCare
+  },
+  {
+    name: 'Dental Exam & X-Rays',
+    img: dentalExamXRays
+  },
+  {
+    name: 'Hygiene Services & Dental Cleaning',
+    img: hygieneServicesDentalCleaning
+  },
+  {
+    name: 'Dental Implants',
+    img: dentalImplants
+  },
+  {
+    name: 'Tooth Fillings',
+    img: toothFillings
+  },
+  {
+    name: 'Invisalign®',
+    img: invisalign
+  },
+  {
+    name: 'Partial & Full Dentures',
+    img: partialFullDentures
+  },
+  {
+    name: 'Root Canals',
+    img: rootCanals
+  }
+];
